+++ src/pages/Offer/sections/Terms.tsx
@@ -2,6 +2,7 @@ import { colors, fonts } from '@hedviginsurance/brand'
 import { TranslationsConsumer } from '@hedviginsurance/textkeyfy'
 import * as React from 'react'
 import styled from 'react-emotion'
+import { getInsuranceTextKey, getPrebuyPDFTextKey } from 'utils/getPDFTextKey'
 import { InsuranceType } from 'utils/insuranceDomainUtils'
 import { HeaderWrapper } from '../components/HeaderWrapper'
 
@@ -71,33 +72,6 @@ const PerilTitle = styled('div')({
   width: '141px',
 })
 
-const getPrebuyPDFTextKey = (insuranceType: InsuranceType): string => {
-  const map = {
-    [InsuranceType.RENT]: 'TERMS_PDF_PREBUY_RENT_URL',
-    [InsuranceType.BRF]: 'TERMS_PDF_PREBUY_BRF_URL',
-    [InsuranceType.STUDENT_RENT]: 'TERMS_PDF_PREBUY_STUDENT_RENT_URL',
-    [InsuranceType.STUDENT_BRF]: 'TERMS_PDF_PREBUY_STUDENT_BRF_URL',
-  }
-
-  if (!map[insuranceType]) {
-    throw new Error(`Invalid insurance type ${insuranceType}`)
-  }
-  return map[insuranceType]
-}
-
-const getInsuranceTextKey = (insuranceType: InsuranceType): string => {
-  const map = {
-    [InsuranceType.RENT]: 'TERMS_PDF_INSURANCE_RENT_URL',
-    [InsuranceType.BRF]: 'TERMS_PDF_INSURANCE_BRF_URL',
-    [InsuranceType.STUDENT_RENT]: 'TERMS_PDF_INSURANCE_STUDENT_RENT_URL',
-    [InsuranceType.STUDENT_BRF]: 'TERMS_PDF_INSURANCE_STUDENT_BRF_URL',
-  }
-
-  if (!map[insuranceType]) {
-    throw new Error(`Invalid insurance type ${insuranceType}`)
-  }
-  return map[insuranceType]
-}
 export const Terms: React.SFC<TermsProps> = ({ insuranceType }) => (
   <Card>
     <HeaderWrapper>
